+++ audiosource_win_pkg/__init__.py
@@ -1,3 +1,3 @@
 """AudioSource Win package."""
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"